+++ src/stages/stage1_events.h
@@ -1,0 +1,13 @@
+#ifndef STAGE1_EVENTS
+#define STAGE1_EVENTS
+
+#include "boss.h"
+
+void cirno_perfect_freeze(Boss*, int);
+void cirno_crystal_rain(Boss*, int);
+void cirno_icicle_fall(Boss*, int);
+void cirno_pfreeze_bg(Boss*, int);
+
+void stage1_events(void);
+
+#endif
