+++ src/stages/stage3.c
@@ -102,6 +102,7 @@ void stage3_fog(int fbonum) {
 	glUniform1f(uniloc(shader, "start"), 0.2);
 	glUniform1f(uniloc(shader, "end"), 0.8);
 	glUniform1f(uniloc(shader, "exponent"), stgstate.fog_exp + 0.5 * sin(global.frames / 50.0));
+	glUniform1f(uniloc(shader, "sphereness"),0);
 	glActiveTexture(GL_TEXTURE0 + 2);
 	glBindTexture(GL_TEXTURE_2D, resources.fbg[fbonum].depth);
 	glActiveTexture(GL_TEXTURE0);
