+++ src/stages/stage4.c
@@ -28,6 +28,7 @@ void stage4_fog(int fbonum) {
 	glUniform1f(uniloc(shader, "start"),0.4);
 	glUniform1f(uniloc(shader, "end"),0.8);
 	glUniform1f(uniloc(shader, "exponent"),4.0);
+	glUniform1f(uniloc(shader, "sphereness"),0);
 	glActiveTexture(GL_TEXTURE0 + 2);
 	glBindTexture(GL_TEXTURE_2D, resources.fbg[fbonum].depth);
 	glActiveTexture(GL_TEXTURE0);
